+++ CHANGELOG.md
@@ -9,3 +9,4 @@ v0.1: June 15, 2018
 - Clean up tab headings to be more consistent
 - Tweak styles on live now button
 - add pull-to-refresh support on the home screen
+- add material bottom tab bar